Close client on unexpected frame error
(For frame errors not associated with a known request)

This is using the standard library's http2 library, so if there is a frame
error, something unexpectedly went wrong on this client or on the remote
server. This shouldn't happen in normal circumstances.

Try to disconnect and establish a new connection

Fixes https://github.com/parse-community/node-apn/issues/64