Proposal: Implement a round-robin MultiClient
It is possible for some use cases to run into limits on the number of
in-flight requests to APNs at a time, especially when sending out bursts
of push notifications.

It seems like this is only necessary when sending out over a hundred
pushes per second, though that may depend on the burstiness, cpu speed or
internet reliability.

Closes https://github.com/parse-community/node-apn/issues/34

Proposal: 5.0.0: Drop support for node < 12

Closes https://github.com/parse-community/node-apn/pull/66

Note that a final 4.x patch release should be tagged first.